+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 1.0.0
+## 1.0.0A
 
 - Initial version.
+
+
+## 1.0.0B
+
+- Process OCValidate output and also show the amount of schema issues at the bottom summary.
+- Now detects OC-EFI-Maker as an invalid configuration (error).
+- Improved Readme.